6. Lottery Refactor
