+++ src/main/java/edu/corvinus/springBootLottery/SpringBootLotteryApplication.java
@@ -1,14 +1,14 @@
-package edu.corvinus.springBootChess;
+package edu.corvinus.springBootLottery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class})
-public class SpringBootChessApplication {
+public class SpringBootLotteryApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringBootChessApplication.class, args);
+		SpringApplication.run(SpringBootLotteryApplication.class, args);
 	}
 
 }
