--- src/main/resources/static/style/Game.css
@@ -1,77 +1,0 @@
-
-body {
-    padding-top: 2vh;
-    text-align: center;
-    height: 100vh;
-    background: linear-gradient(#b92b27, #1565c0);
-}
-
-.c-piece {
-    width: 80px;
-    height: 80px;
-    display: inline-block;
-    text-align: center;
-    line-height: 82px;
-    font-size: 40px;
-    font-family: monospace;
-    cursor: pointer;
-    vertical-align: middle;
-}
-
-.c-piece img{
-    vertical-align: middle;
-}
-
-.c-row {
-    padding: 0px;
-    margin: 0px;
-}
-
-
-.dark-brown-background {
-    background-color: #CC5803;
-}
-
-.light-brown-background {
-    background-color: #FFC15E;
-}
-
-#chessBoard {
-    margin: 0 auto;
-    width: 640px;
-    height: 640px;
-    border: solid #1F1300 20px;
-}
-
-#submitMove {
-    background-color: rgb(38, 255, 74);
-    border: none;
-    margin-top: 10px;
-    margin-right: 10px;
-    width: 300px;
-    height: 50px;
-    font-size: 30px;
-    cursor: pointer;
-}
-
-#cancelMove {
-    background-color: rgb(255, 0, 0);
-    border: none;
-    margin-top: 10px;
-    margin-left: 10px;
-    width: 300px;
-    height: 50px;
-    font-size: 30px;
-    cursor: pointer;
-}
-
-.highlight-this {
-    background-color: red;
-}
-
-#topbar h1 {
-  margin-left: 40px;
-  margin-right: 40px;
-  display: inline-block;
-  font-family: Helvetica;
-}
