--- src/main/resources/templates/Error.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="utf-8">
-        <title>Error</title>
-    </head>
-    <body>
-        <h1>Default error page for website.</h1>
-    </body>
-</html>
