--- src/main/resources/templates/Game.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="utf-8">
-        <title>Game Page</title>
-
-        <link rel="stylesheet" type="text/css" href="/chess/static/style/Game.css">
-        
-        <script src="/resources/static/js/jquery-3.6.4.min.js"></script>
-        <script src="/resources/static/js/Game.js"></script>
-
-    </head>
-    <body>
-        <div id="topbar">
-            <h1 id="chessId" th:text="${game.id}">Not Found</h1>
-            <h1 id="gameTurn"></h1>
-        </div>
-        <div id="chessBoard"></div>
-    </body>
-</html>
