--- src/main/resources/templates/Home.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="utf-8">
-        <title>Home Page of Testing Site by MAXIMILIAN</title>
-    </head>
-    <body>
-        <h1>Chess Site</h1>
-
-        <a href="/chess/makeNewChess">
-            <h2>New Game</h2>
-        </a>
-
-        <a href="/chess/getChessById">
-            <h2>Get Chess Game by ID</h2>
-        </a>
-
-    </body>
-</html>
