--- src/main/resources/templates/MakeGame.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="utf-8">
-        <title>Game Page</title>
-    </head>
-    <body>
-        <form action="/chess/makeNewChess" method="POST">
-            <label for="submitNewChess">New Game: </label>
-            <input type="submit" name="submitNewChess">
-        </form>
-    </body>
-</html>
