--- src/main/resources/templates/chess.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:th="http://www.thymeleaf.org">
-    <head>
-        <meta charset="utf-8">
-        <title>Chess Page of Testing Site by MAXIMILIAN</title>
-        <script src="/resources/static/js/jquery-3.6.4.min.js"></script>
-        <script src="/resources/static/js/chess.js"></script>
-    </head>
-    <body>
-        <p>Html loads</p>
-    </body>
-</html>
