+++ src/test/java/edu/corvinus/springBootLottery/SpringBootChessApplicationTests.java
@@ -1,4 +1,4 @@
-package edu.corvinus.springBootChess;
+package edu.corvinus.springBootLottery;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
